Extracting the sentEther fucntionality
